requirements.txt
aniso8601==7.0.0
asgiref==3.2.5
autopep8==1.5
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-graphql-jwt==0.3.0
django-heroku==0.3.1
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.1
gunicorn==20.0.4
promise==2.3
psycopg2-binary==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
